+++ src/main/java/nl/lexemmens/podman/service/PodmanExecutorService.java
@@ -14,6 +14,7 @@ import org.zeroturnaround.exec.stream.slf4j.Slf4jStream;
 import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.regex.Pattern;
 
 import static nl.lexemmens.podman.enumeration.TlsVerify.NOT_SPECIFIED;
 
@@ -170,7 +171,7 @@ public class PodmanExecutorService {
         } catch (MojoExecutionException e) {
             // When the command fails, the whole command is put in the error message, possibly exposing passwords.
             // Therefore we catch the exception, remove the password and throw a new exception with an updated message.
-            String message = e.getMessage().replaceAll(String.format("-p[, ]+%s", password), "-p **********");
+            String message = e.getMessage().replaceAll(String.format("-p[, ]+%s", Pattern.quote(password)), "-p **********");
             log.error(message);
             throw new MojoExecutionException(message);
         }