+++ src/test/java/nl/lexemmens/podman/service/PodmanExecutorServiceTest.java
@@ -115,6 +115,22 @@ public class PodmanExecutorServiceTest {
     }
 
     @Test
+    public void testLoginPasswordObfuscatedWithComplexPasswordUponFailure() throws MojoExecutionException {
+        PodmanConfiguration podmanConfig = new TestPodmanConfigurationBuilder().setTlsVerify(FALSE).initAndValidate(mavenProject, log).build();
+
+        podmanExecutorService = new PodmanExecutorService(log, podmanConfig, commandExecutorDelegate);
+
+        when(commandExecutorDelegate.executeCommand(isA(ProcessExecutor.class))).thenThrow(new MojoExecutionException("Command failed: podman login --tls-verify=false registry.example.com -u username -p df8&35%)!AL"));
+
+        try {
+            podmanExecutorService.login("registry.example.com", "username", "df8&35%)!AL");
+            Assertions.fail("This should not happen");
+        } catch (MojoExecutionException e) {
+            Assertions.assertEquals("Command failed: podman login --tls-verify=false registry.example.com -u username -p **********", e.getMessage());
+        }
+    }
+
+    @Test
     public void testPush() throws MojoExecutionException {
         PodmanConfiguration podmanConfig = new TestPodmanConfigurationBuilder().setTlsVerify(TRUE).initAndValidate(mavenProject, log).build();